Add register to vote CTA to single proposal view
